+++ packages/searchkit-apollo-resolvers/src/index.ts
@@ -1,4 +1,0 @@
-export * from './resolvers'
-export * from './query'
-export * from './facets'
-export { default as SearchkitSchema } from './schema'