+++ packages/searchkit-apollo-resolvers/src/resolvers/index.ts
@@ -1,5 +1,0 @@
-export { default as HitsResolver } from './HitsResolver'
-export { default as ResultsResolver, SearchkitConfig } from './ResultsResolver'
-export { default as FacetsResolver } from './FacetsResolver'
-export { default as SummaryResolver } from './SummaryResolver'
-export { default as SearchkitResolvers } from './SearchkitResolvers'