+++ packages/searchkit-schema/src/index.ts
@@ -1,0 +1,4 @@
+export * from './resolvers'
+export * from './query'
+export * from './facets'
+export { default as SearchkitSchema, SearchkitSchemaConfig } from './schema'