+++ packages/searchkit-schema/src/resolvers/index.ts
@@ -1,0 +1,4 @@
+export { default as HitsResolver } from './HitsResolver'
+export { default as SearchkitResolver, SearchkitConfig } from './ResultsResolver'
+export { default as FacetsResolver } from './FacetsResolver'
+export { default as SummaryResolver } from './SummaryResolver'